management/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.28.0
docker>=7.0.0
pydantic-settings>=2.7.0
python-multipart>=0.0.18
aiofiles>=24.1.0
websockets>=13.0
